+++ cmd/zdb/Makefile.am
@@ -16,5 +16,4 @@ zdb_SOURCES = \
 
 zdb_LDADD = \
 	$(top_builddir)/lib/libnvpair/libnvpair.la \
-	$(top_builddir)/lib/libzfs/libzfs.la \
 	$(top_builddir)/lib/libzpool/libzpool.la
