+++ cmd/zhack/Makefile.am
@@ -11,5 +11,4 @@ zhack_SOURCES = \
 
 zhack_LDADD = \
 	$(top_builddir)/lib/libnvpair/libnvpair.la \
-	$(top_builddir)/lib/libzfs/libzfs.la \
 	$(top_builddir)/lib/libzpool/libzpool.la
