+++ cmd/zinject/Makefile.am
@@ -13,5 +13,4 @@ zinject_SOURCES = \
 
 zinject_LDADD = \
 	$(top_builddir)/lib/libnvpair/libnvpair.la \
-	$(top_builddir)/lib/libzfs/libzfs.la \
-	$(top_builddir)/lib/libzpool/libzpool.la
+	$(top_builddir)/lib/libzfs/libzfs.la
