+++ cmd/ztest/Makefile.am
@@ -20,7 +20,6 @@ ztest_SOURCES = \
 
 ztest_LDADD = \
 	$(top_builddir)/lib/libnvpair/libnvpair.la \
-	$(top_builddir)/lib/libzfs/libzfs.la \
 	$(top_builddir)/lib/libzpool/libzpool.la
 
 ztest_LDADD += -lm
