+++ lib/libzfs/libzfs_iter.c
@@ -33,6 +33,7 @@
 #include <stddef.h>
 #include <libintl.h>
 #include <libzfs.h>
+#include <libzutil.h>
 #include <sys/mntent.h>
 
 #include "libzfs_impl.h"
