+++ lib/libzfs/libzfs_sendrecv.c
@@ -54,6 +54,7 @@
 
 #include <libzfs.h>
 #include <libzfs_core.h>
+#include <libzutil.h>
 
 #include "zfs_namecheck.h"
 #include "zfs_prop.h"
